% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_2cop.R
\name{test_2cop}
\alias{test_2cop}
\title{Compute Similarity Between Two Copulas}
\usage{
test_2cop(x, y, Nsim = 100, paired = FALSE, alpha = 0.95)
}
\arguments{
\item{x}{Numeric matrix representing the first sample of copula data.}

\item{y}{Numeric matrix representing the second sample of copula data.}

\item{Nsim}{Integer specifying the number of simulations to perform (default is 100).}

\item{paired}{Logical indicating whether the samples are paired (default is FALSE).}

\item{alpha}{Numeric value indicating the confidence level for the test (default is 0.95).}
}
\value{
A list containing:
  - `pvalue`: the p-value of the test indicating similarity.
  - `cvm`: the observed test statistic.
  - `VaR`: the critical value at the specified confidence level.
  - `cvmsim`: the simulated test statistics.
}
\description{
This function performs a statistical test to evaluate the similarity between
two copulas represented by the datasets `x` and `y`. It supports both paired
and unpaired samples and utilizes Monte Carlo simulation to estimate the
p-value and critical value of the test statistic.
}
\examples{
x = matrix(runif(100), ncol=2)
y = matrix(runif(100), ncol=2)
result <- test_2cop(x, y, Nsim=1000, paired=FALSE, alpha=0.95)
print(result)

}
