\name{TwoCop-package}
\alias{TwoCop-package}
\docType{package}
\title{Nonparametric test of equality between two copulas
}
\description{This package implements the nonparametric test of equality
between two copulas proposed by Remillard and Scaillet (2009) in their JMVA
paper. The test is based on the Cramer-von-Mises statistic between the two
empirical copulas. An approximate p-value is returned. 
}
\details{
\tabular{ll}{
Package: \tab TwoCop\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-17\cr
License: \tab GPL-2\cr
}
The function \code{TwoCop} provides an approximate p-value for the test of equality
between two copulas. 
}
\author{Bruno Remillard and Jean-Francois Plante


Maintainer: Bruno Remillard <bruno.remillard@hec.ca>
}
\references{
Remillard, B. & Scaillet, O. (2009) Testing for equality between two copulas. Journal of Multivariate Analysis, 100, 377-386.
}

\keyword{ package }
\keyword{ nonparametric }
\keyword{ multivariate }
\keyword{ htest }
\keyword{ copula }
\keyword{ ranks }
